rename loss to metrics
